[bugfix] bean을 못찾고 있음
Consider defining a bean of type 'org.springframework.kafka.core.KafkaTemplate' in your configuration.
