+++ docker/docker-compose.yml
@@ -44,9 +44,9 @@ services:
     hostname: kafka
     image: confluentinc/cp-kafka:5.5.1
     ports:
-      - "9095:9095"
+      - "9092:9092"
     environment:
-      KAFKA_ADVERTISED_LISTENERS: LISTENER_DOCKER_INTERNAL://kafka:19095,LISTENER_DOCKER_EXTERNAL://${DOCKER_HOST_IP:-127.0.0.1}:9095
+      KAFKA_ADVERTISED_LISTENERS: LISTENER_DOCKER_INTERNAL://kafka:19092,LISTENER_DOCKER_EXTERNAL://${DOCKER_HOST_IP:-127.0.0.1}:9092
       KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: LISTENER_DOCKER_INTERNAL:PLAINTEXT,LISTENER_DOCKER_EXTERNAL:PLAINTEXT
       KAFKA_INTER_BROKER_LISTENER_NAME: LISTENER_DOCKER_INTERNAL
       #KAFKA_ADVERTISED_HOST_NAME: host.docker.internal
@@ -64,7 +64,7 @@ services:
     ports:
       - "9001:9001"
     environment:
-      KAFKA_BROKERCONNECT: "kafka:19095"
+      KAFKA_BROKERCONNECT: "kafka:19092"
       JVM_OPTS: "-Xms16M -Xmx48M -Xss180K -XX:-TieredCompilation -XX:+UseStringDeduplication -noverify"
       SERVER_PORT: 9001
       MANAGEMENT_SERVER_PORT: 9001
